fix: review fixes for the eBPF backend runtime
Sonnet review of 371da2c returned MERGE-READY with two Medium findings and
one Low; all three fixed here.

- vmlinux.h declared usb_bus.busnum / usb_device.devnum as __u16/__u8, but
  the kernel defines both as `int` (verified against this host's
  /sys/kernel/btf/vmlinux). BPF_CORE_READ reads sizeof(declared type) bytes
  at the CO-RE-relocated offset, so the narrow declarations read only the
  low bytes -- correct only by little-endian accident, wrong on big-endian.
  Declare both as `int`; the BPF program narrows the read *value* into
  key_t's u16/u8 afterward (endian-safe). Comment both sites so nobody
  "simplifies" the widths back.

- traffic_delta() narrowed the u16 map busnum to bus_id with `as u8`, which
  wraps an out-of-range bus onto some other real bus's low byte. Use
  u8::try_from(..).unwrap_or(0), matching binary.rs's own busnum handling
  (fall back to an obviously-unknown bus 0). Added a test mirroring
  binary.rs's oversized_busnum_falls_back_to_zero.

- The eBPF poller swallowed a disconnected delta receiver and kept polling
  until stop(); now it sets the shutdown flag on Disconnected so it self-
  terminates, matching run_source_chain's packet readers. run holds only a
  shared &borrow of the same flag, so setting it from the send closure is
  sound.

Gates green both configs (449/453 unchanged) and --features ebpf (470 unit,
+1 for the new fallback test). Re-verified live after the vmlinux.h change:
the eBPF backend still byte-matches ground truth (0.935 MB/s vs 0.825 MB/s
MJPEG payload = 1.13x, the expected UVC/iso framing overhead), with the
camera still correctly attributed to bus 1 dev 4.

+++ src/bpf/vmlinux.h
@@ -10,9 +10,17 @@
  * kernel produced it). It declares only the structs and fields the BPF
  * program actually reads. CO-RE (`BPF_CORE_READ` / `preserve_access_index`)
  * relocates every access against the *target* kernel's BTF at load time by
- * field name, not by the offsets implied here, so the field layout below
- * does not need to match the real kernel struct layout -- only the field
- * names and types need to match.
+ * field name, so the field *offsets* implied by the layout below need not
+ * match the real kernel struct layout.
+ *
+ * Each field's declared *width* still must, though: `BPF_CORE_READ` reads
+ * exactly `sizeof(the locally declared type)` bytes at the relocated offset.
+ * Declaring a field narrower than the kernel's real field (e.g. `__u16` for
+ * a kernel `int`) reads only its low bytes -- correct only by little-endian
+ * accident, wrong on a big-endian target. So `busnum`/`devnum` are `int`
+ * here because the kernel defines them as `int` (verified against this
+ * host's `/sys/kernel/btf/vmlinux`); the BPF program narrows the read *value*
+ * into `key_t`'s `__u16`/`__u8` fields afterward, which is endian-safe.
  */
 
 typedef signed char __s8;
@@ -129,12 +137,12 @@ struct usb_host_endpoint {
 };
 
 struct usb_bus {
-	__u16 busnum;
+	int busnum;
 };
 
 struct usb_device {
 	struct usb_bus *bus;
-	__u8 devnum;
+	int devnum;
 };
 
 struct urb {
